Rename theme to 'circuit'
