+++ src/components/Col/Col.story.js
@@ -4,14 +4,14 @@ import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 import { GROUPS } from '../../../.storybook/hierarchySeparators';
 
-import { standard } from '../../themes';
+import { circuit } from '../../themes';
 
 import withTests from '../../util/withTests';
 import Col from './Col';
 
 const StyledCol = styled(Col)`
-  background-color: ${standard.colors.b500};
-  color: ${standard.colors.white};
+  background-color: ${circuit.colors.b500};
+  color: ${circuit.colors.white};
   font-size: 14px;
   font-weight: bold;
   line-height: 20px;
